Implement LazyRow-like rendering of visible screens in Pager

+++ app/compendium/src/main/java/cz/frantisekmasa/wfrp_master/compendium/ui/CompendiumScreen.kt
@@ -73,7 +73,7 @@ private fun TopBar(routing: Routing<Route.Compendium>) {
 
 @Composable
 private fun BoxWithConstraintsScope.MainContent(routing: Routing<Route.Compendium>) {
-    val screenWidthPx = constraints.maxWidth.toFloat()
+    val screenWidthPx = constraints.maxWidth
     val screenWidthDp = maxWidth
 
     Column {
